remove legacy_utils/
