+++ README.md
@@ -33,8 +33,7 @@ pip install -e .
 
 ```python
 import os
-from ndx_miniscope.legacy_utils import read_settings, read_notes, read_miniscope_timestamps
-from ndx_miniscope.utils import get_starting_frames
+from ndx_miniscope.utils import get_starting_frames, read_settings, read_notes, read_miniscope_timestamps
 from pynwb import NWBFile, NWBHDF5IO
 from datetime import datetime
 from dateutil.tz import tzlocal
