--- src/pynwb/ndx_miniscope/legacy_utils/__init__.py
@@ -1,1 +1,0 @@
-from .utils import *
